requirements.txt
Django==5.2.4
python-dotenv
Pillow
django-jazzmin==3.0.1
whitenoise
chardet
dbfread
gunicorn